// src/tcgcommerce/modules/inventory/product/card/service/import/job/provider/dto/inventory.product.card.service.import.job.provider.dto.ts
export class InventoryProductCardServiceImportJobProviderDTO {
    inventoryProductCardServiceImportJobProviderTCGPlayerId: number;
    inventoryProductCardServiceImportJobProviderSetName: string;
    inventoryProductCardServiceImportJobProviderProductName: string;
    inventoryProductCardServiceImportJobProviderNumber: string;
    inventoryProductCardServiceImportJobProviderRarity: string;
    inventoryProductCardServiceImportJobProviderCondition: string;
    inventoryProductCardServiceImportJobProviderPrinting: string;
    inventoryProductCardServiceImportJobProviderQty: number;
}