// src/tcgcommerce/modules/inventory/product/card/service/import/job/provider/type/interface/inventory.product.card.service.import.job.provider.type.interface.ts
export interface InventoryProductCardServiceImportJobProviderTypeDataKey {
    inventoryProductCardServiceImportJobProviderTypeDataKeyTCGPlayerId: string;
    inventoryProductCardServiceImportJobProviderTypeDataKeyCondition: string;
    inventoryProductCardServiceImportJobProviderTypeDataKeyPrinting: string;
    inventoryProductCardServiceImportJobProviderTypeDataKeyQty: string;
}